Run the docker build workflow for each pull request
We run our Ruby CI for each pull request, but we only run our Docker
build workflow for the main branch.

Sometimes, a PR is merged and breaks the Docker build workflow. It's
important to see if it's failing before merging the PR.

At first, we didn't run it for each pull request because it took too
long.

But with the help of caching, it's now fast enough to run for each pull
request.