Update the PG version to 15.8 in the CI workflows
We are already using 15.8 in development as specified in
`.tool_versions`. It's a good practice to keep the same version across
all environments, including CI.

By the way, production is using 15.10, so we might upgrade development
to 15.10 as well. I’m not doing it now because I don’t want to break
anyone’s local development environments.