Switch period position in multi-line expressions to "trailing"
This is particularly dire in `pry` where the idiomatic/default `.`
starts a runs a command, e.g. `.ls` runs the `ls` command.  Using
trailing dots solves this problem, it's also the common code style in
Jeremy Evans software libraries.

As an additional complication, also use a new, patched `erb-formatter`
that does not rewrite ruby code redundant to rubocop, because, it
writes them with leading periods using its own code formatter (a basic
use of `syntax_tree`) that is not easily configured.