Fix race condition on saving CNI datastore file
When a pod is created, kubelet will call the CRI to setup the container
and its networking and CRI might call the CNI in parallel multiple times.

The code in CNI had no locking mechanism for saving/reading the file and as
a result, we could see some inconsistencies across cni datastore file.

This commit fixes the issue by adding locks to the files and protecting the
parts which access the file