Switch PostgreSQL resource creation to using forme
The PostgreSQL creation form is much more complicated than the
firewall creation form, and serves as a better test of whether using
forme is a better approach in general.

After working on this, I'm of the opinion that most of the complexity
in the current resource creation form approach is inherent and not
incidental, meaning that there doesn't exist an approach that can
avoid most of the complexity.  It is, by its nature, a complex
problem.

Overall, I think this approach is simpler, but not much simpler, and
it may only be simpler to me as I have a good understanding of forme.

Some advantages of this approach compared to the resource creation
form approach:

* No inline javascript is required

* The javascript is about half the size, and is more declarative
  in nature (the resource creation form javascript is more imperative
  in nature).  One change is the new javascript hides rather than
  disables inputs.  I think this fixes certain issues, but it also
  requires some test changes to choose the options, since the
  generated HTML does not have options automatically disabled.

* Certain things are handled automatically by Forme, such as CSRF
  tags, selecting/checking current defaults for model inputs.

The main disadvantage I think is that if you do need to modify the
HTML, instead of editing an ERB template, you would edit the Ruby
code in clover_forme_config.rb.  The Ruby code is more verbose than
the ERB code.  Forme was designed for cases where you would set an
id/class attribute on the form or specific inputs, and style
appropriately based on that.  It wasn't designed for the tailwind
case where you have 5 layers of nested tags, with a class attribute
at most/every level.  Forme can certainly work with it, but it takes
significantly more customization code.

Based on feedback, the PostgreSQL form reuses more of the existing
code.  It modifies the OptionTreeGenerator to output information for
Forme to use.  It uses the ContentGenerator to generate content for
some of the more advanced radio buttons.

This inlines the partnership notice, since the PostgreSQL form was
the only form using it, and it seemed easier than rebuilding it
in Forme.

This adds details to Sequel::ValidationFailed error messages in the
API.  In some cases, doing so is redundant with the main error
message.  You see an example of this in the tests.  Maybe we should
make the Sequel::ValidationFailed and Validation::ValidationFailed
error display more similar?  I tried that, but it resulted in ~10
test failures, so that isn't included in this commit.

This only moves some of the validations from the postgres_post
method into the PostgresResource model.  We could move more, but
this commit is large as-is.

This removes the validate_request_params call.  I don't think it
is necessary.  In the web case, it is fine to ignore additional
parameters.  In the api case, unrecognized parameters should be
rejected, but they are already rejected by committee, so the
checking in the route is redundant.