Read replica lag detection adn server recycling
Even if we setup replication, it is possible that the replication would
fail and we might need to recycle the server. To detect that, we are
continuously checking the timeline_id difference between a read replica
and its parent. If the timeline_id difference is in place in another 2
minutes, then we decide to recycle the read replica.

We also implement the read replica failover mechanism in this commit
which is essentially refreshing dns records marking the new server as
the representative.