Explicitly set node-ip in kubeadm config using kubelet extra args
Kubelet determines the ips peresent on the node and since we have both
ipv4 and ipv6 ips, we expect it to find the ips but it acts in an
underministic way, sometimes finding the ip and sometimes no.

We will explicitly provide the ipv4 and ipv6 ips to the each cluster
vm using the kubelet extra args API present in the kubeadm config file.

Also the API version of kubeadm is upgraded from v1beta3 to v1beta4.

Join scripts for worker and control-plane nodes are also merged and
with the help of a flag, the script is reused.