Handle display_state generation during resource deletion
During the resource deletion, it is possible to have a short time where the
strand is nil, which causes the display_state to throw and exception. This
commits fixes that by moving delete state check to the beginning. It also
refactors the display_state tests, because over the time display state got
move complex and testing all possible display_state outputs using a single
test is not very feasible anymore.