Drop the subnet when a k8s cluster is destroyed
As a bug from the pre-CCM times, we forgot dropping the subnet that's
created internally when a k8s cluster is created. This patch adds the
incr_destroy call to the destroy label of KubernetesClusterNexus.

Theoretically, a Kubernetes cluster can share a subnet with other resources
when an existing subnet is supplied to KubernetesClusterNexus.assemble.
However, this is not used in the mainstream routes so we are not worried about
the case.