Read replica lag detection and server recycling
Even if we setup replication, it is possible that the replication would
fail and we might need to recycle the server. To detect that, we are
continuously checking the lag between a read replica and its parent. If
the applied wal difference is greater than 80 MB and the replica does
not apply new wal files in the upcoming 15 minutes, we decide to recycle
the server.
We also implement the read replica failover mechanism in this commit
which is essentially refreshing dns records marking the new server as
the representative.