Postgres read replicas web and API support
With this commit, read replicas are available in the UI. We can
provision a read replica basically by entering a name and clicking to
the create button. We are not allowing customers to pick different spec
for read replicas because it opens up a bunch of options for scale
operations.