Prioritize AX102 hosts for premium CPU runner testers
We will introduce a new runner tier soon; I'm working on a PR to enable
it.

We have some AX102 machines in our fleet. Until the new runner tier is
launched, we want to give higher priority to some customers to test
them.

We get the runner object in the VM nexus to check if it’s a preferred
performance CPU. It’s a bit of an antipattern, but making changes to the
codebase would require a lot of effort. We need to pass the preference
from runner nexus to VM nexus to allocator. Additionally, it allocates
the VM at the `start` label, not in `assemble`, so we need to persist
the preference in the stack. To avoid all of this, I just try to get the
runner if the location is "github-runners." This code will be removed
after the new runner tier is launched.

Since we don't have a VM pool for them, we don't pick from the pool if
it's a preferred performance CPU.