Introduce a new "performance" vm family
We're working on a new runner tier that has higher CPU performance. I've
introduced a new vm family to use with this tier. The VMs with better
CPU performance belong to this family.

I added a billing rate only for the "github-runners" location since we
don't introduce performance VMs publicly that are only used internally
for premium runners. For the VmVCpu billing rate, I simply doubled the
standard price. Since we don't create billing records for internal
projects, the value is not important at all for the runner location. We
can adjust the value for other locations when we decide to make the new
family available to other services.