Add a family filter to the VM host allocator
It selects only VM hosts for the specified families. If the premium CPU
runners are enabled, it tries to allocate for both standard and
performance families, giving higher priority to the performance family
if there is capacity. Otherwise, it falls back to the standard family.

Since we use the same hosts for standard and burstable virtual machines,
we pass it as "standard" for "burstable" instances.