Allow Prog::Vm::Nexus.assemble to choose cloud-hypervisor version
Additionally, allow it to choose whether to support hugepages and to
choose the firmware version (though currently only one firmware
version is supported).  The default for hugepages is set to "on,"
pending additional work to avoid the reboot overheads of hugepages in
development.

In rhizome prep_host, download all supported cloud-hypervisor versions
and firmware versions.  Currently, only one firmware version is
supported.  On Ubuntu 24, both cloud-hypervisor 35.1 and 45.0 are
installed, with 45.0 as the default if it is installed.  On Ubuntu
22, only cloud-hypervisor 35.1 is installed.