Restart node drain on failure
kubectl drain node_name is an idempotent command which allows us to
easily retry it

During the upgrade, since a new node has just joined the cluster, the etcd
leader might change and the control plane nodes are not stable. That can
easily disrupt the drain command but we will keep retrying to drain the node

Since a 1hour deadline is set on the drain node, we know we will get the page
and we won't restart indefinitely.