Change Allocator.Request from Struct to Data
I found that Struct doesn't support named parameters. A long list of
unnamed parameters makes it difficult to read.

Additionally, I plan to make some of the allocator parameters more
flexible and pass them as a hash, which named parameters help.

Since Data is an immutable version of Struct with named parameters, I
made the change. It also requires passing all properties.

I applied the Data changes in the following commits for clarity.