Set default values for `Allocator.Request` in one place
We set default values for some parameters in `self.allocate` and others
in the constructor. I moved all default values to the constructor to
make them clearer.

Since Data is immutable, we can't change its values directly in tests.
So, we need to create a new request object for some tests.