Pass allocator options as a single parameter
We add an allocator property for new features and pass it as a separate
parameter to Nexus assemble. This complicates Nexus. Additionally, we
introduce new branching for the new features. If we combine all
parameters into a single hash and allow passing any property of the
allocator, we can shift the responsibility of tuning allocator
preferences to the caller. This simplifies Nexus.assemble and allows us
to keep special logic in the responsible progs.

We can also write any special allocator combination for any VM in
production.