Show only the company name on the invoice if it's provided
In the past, we didn't have a separate field for the company name, so
customers used the billing name as the company name.

Now that we've added a company name field, it appears on the invoice
twice. In addition, some customers have requested that only the company
name be shown on the invoice, not the person's name.

Therefore, if the company name is provided, there's no need to include
the person's name.

We didn't do this before because we were generating invoice PDFs on the
fly, which would have altered already generated invoices. However, we
have started to persist the invoices in blob storage and display them
from there, so these changes won't affect old invoices.