Update bin/monitor to export metrics to VictoriaMetrics
This commit updates bin/monitor to scan for resources which have metrics
export capability (currently only PG servers) and periodically export
metrics out and store them in VictoriaMetrics.

Each resource which exports metrics should include `MetricsTargetMethods`,
which contain the config and a commmon algorithm to fetch scraped data
from a given list of endpoints. Roghly the logic looks like this:

Producer side (PG server):
A systemd-backed timer executes `postgres/bin/metrics-collector`, which
scrapes the list of endpoints and stores them on disk, every 15s
(default interval). This location is the `pending` buffer, which can
grow upto a 120 entries by default, providing recovery for 30
minutes of outages in other parts of the system.

Consumer side (bin/monitor):
Monitor periodically connects to each metrics target resource to fetch
upto 4 scrapes at a time and write them to VictoriaMetrics, and move the
successfully exported scrapes to the `done` buffer, which is cleand up
periodically by the `metrics-collector`.  Currently, we only create a
single global VictoriaMetrics instance for simplicity, although it can
be configured to be location-aware.