Fix missed query parameterization in specs
After these changes, the query parameterization log shows 0 queries
with missed parameters.

None of these usages are unsafe, since they are only in the specs
and use trusted data, but direct interpolation patterns such as
`'#{r1.id}'` are a bad idea in the specs, as it gives the impression
to developers that such constructions are safe to use elsewhere in
the application.

Bad examples anywhere are a threat to safe usage everywhere.