Add Clover#check_found_object to DRY up code
This DRYs up code in many routes.  Additionally, this fixes cases
where Clover would previous return 204 results for DELETE requests
with garbage at the end of the path.  204 should only be used if
the path would actually be valid.  This checks that the path has
been fully routed before using a 204 response.