Check all requests are appropriately authorized when running non-frozen specs
This adds a needs_authorization flag before the main hash branch dispatch.

When running non-frozen specs, if that flag is not cleared before a response
is returned, an exception is raised.  The flag can be cleared by:

* Calling an authorization method
* Calling no_authorization_needed to explicit indicate authorization is not needed

To find incorrect usage of no_authorization_needed, it also raises if the
request is already authorizated or does not require authorization.

This support found the missing postgres read-replica authorization, and should
prevent missing authorization in the future.