Add hugepage, cloud-hypervisor, and firmware fields to params.json
These fields are not yet read by `rhizome`.

The firmware and cloud-hypervisor version fields allow control over
the versions used at provisioning time, enabling gradual rollout and
testing of new versions and combinations. This is the main motivation.
`nil` values for all of these are acceptable, a near-future version of
`rhizome` will apply a default in that case.

The hugepage field is for skipping VM allocation using hugepages, to
speed up some aspects of development iteration: hugepages cost an
extra reboot in `prep_host` and slows down every reboot as Linux
searches for contiguous memory.  Alone, this does not complete the
objective, but it's also necessary to move towards it.

Commit message by Daniel Farina <daniel@ubicloud.com>