Unify GitHub runner allocator logic
When we allocate a VM for a runner, we need to consider different
scenarios, such as allowing US for standard-60 runners or allowing US
runners for specific customers. We currently have separate feature flags
for each scenario, which were added for specific customers. Instead of
adding feature flags for every situation, we can maintain allocator
preferences. This approach offers more flexibility and is easier to
manage, allowing us to tailor it for these specific customers.

For now, we'll skip the VM pool if performance runners are preferred.
We'll implement the pool for them later.

I have two other improvements in mind, but I will address them later:
-  To eliminate the allocator branching in Vm::Nexus, I will move the
   GitHub related parts to the GithubRunner prog and the force_id logic
   to the allocator. This will significantly clean up the code.
-  We also have the same allocator preferences in the VM stack as a
   separate property. With each feature, the number of properties will
   increase. We can integrate allocator preferences as a property of the
   VM stack, making it easier to manage and maintain. Then we can simply
   pass allocator_preferences hash to the Allocator, allowing it to
   override the default preferences. I believe this will also streamline
   the code, especially the tests, since we need to test each branch
   separately.