Use flash and pre_selected_options to set default values
We already use `option_dirty` hash to remember the values that were selected by
users explicitly. This allows us to keep the selected values as selected even
when user changes the parent option. However, we used to don't remember these
values after the form submission and in resource update forms. We are starting
to populate `option_dirty` from `flash` and `pre_selected_options` hash, so
that we can remember user's selections in more cases.