Update the family of the VM if it's upgraded
If there's available capacity and a customer wants to upgrade to a
better CPU, we upgrade their runners. So, even if the VM is initially in
the "standard" family, it might be allocated on a "premium" family host.
We can check the family of the VM host for the effective family.
However, for historical data and utilization calculations, it's better
to keep the data consistent and update the family of the virtual
machine.

We use "standard" family hosts for "burstable" family VMs.