Add Postgres metrics UI
This commit adds the UI components for Postgres metrics, behind a
feature flag. The charts are built on top of Apache ECharts [1].
After some prototyping and reading source code across d3.js, ApexCharts
and ECharts, I decided to pick ECharts as it provides a good balance of
flexibility, performance and out-of-the box features, with a
navigable codebase and documentation.

For this version, I have added the metrics as another card on the
Postgres overview page. But this page is already pretty overcrowded.
In a future version, I would like to split the metrics section to a
separate place to clean things up and also avoid unnecessary requests to
fetch metrics on each Postgres server overview visit.

[1]: https://github.com/apache/echarts/