Remove use of from_ubid in top-level project branch
Include project visibility check in the query.

It would be good to only return projects associated with the
current account, but that would change 403 responses into 404
responses.  I think that is better for security, as 403 leaks
information about whether the project exists, but that change
can be made later.

Specs that mocked the from_ubid call need to be updated to
handle this.