Remove usage of PaymentMethod.from_ubid in billing route
Check the billing_info_id matches as part of the query. Also,
check that the billing_info_id is not NULL.  Previously, if
the project and payment_method both had a NULL billing_info_id,
this check would pass, only to fail with a NoMethodError later.
I'm not sure whether that can actually happen in the application,
but at least both database fields do not have a NOT NULL
constraint.