Fix query-per-object and other issues in load balancer patch route
Previously, this resulted in a query for each requested VM, as
well as a query for the load balancer for each requested VM.  This
changes it to 2 queries, one to get the VMs, and one to get the
load balancers for those VMs.

Previously, if multiple VMs were submitted, and one was authorized
but the second was not, the code would add the one VM that was
authorized, but not the one that wasn't authorized, and it wouldn't
update the load balancer.  Change the transaction scope to handle
all database actions, so either all of them are applied or none
of them are.