Allow creating private subnet with same name as previously destroyed private subnet
This failed because each time you create the private subnet, it tries
to create the -default firewall.  However, destroying the private
subnet does not destroy the related -default firewall.  So the second
creation with the same name would fail when attempting to create the
firewall.

Detect that case before it happens, and if the -default firewall
already exists, append a random string to the name.

This is prone to a race condition if the user concurrently creates
a -default firewall manually, but the worst case scenario in that case
is a bogus error message to the user, which I think is acceptable.
It is possible to use a savepoint and rescue the exception and retry,
but I don't think the benefit of avoiding the race condition is worth
the complexity cost.