Ensure automatically created firewall names are valid
Previously, creating a private subnet with a name longer than 55
characters resulted in a firewall with name over 63 characters,
which the route code would not recognize as valid, and clicking
the links for the firewall would result in a 404.

Avoid the issue by truncating the private subnet name appropriately
to stay under the limit.