Enable GitHub client auto pagination explicitly
When the "auto_paginate" flag is enabled, the Ockokit client will
paginate until there are no resources left.

This can be risky depending on the API. For example, if you’re
retrieving the list of jobs for a workflow, it’s fine since a single run
has a limited number of jobs. However, if you’re getting the list of
runs for repositories, it’s risky because there could be hundreds of
pages.

Auto-pagination is acceptable for API calls in respirate since they are
asynchronous, but it's not ideal to paginate extensively in web
requests.

Therefore, I have disabled it by default and only enable it for clients
in respirate.