Set the GitHub API's per_page value at the client level
The GitHub API returns 30 resources per page by default and allows a
maximum of 100 resources per page.

When the client has auto_paginate enabled, it automatically sets the
per_page value to 100.

It seems configurable at the client level. Instead of setting it for
each request separately, we can set it at the client level.