Avoid sleeping inside transaction in runtime github route
This actually avoids the transaction completely.  Previously,
the behavior was:

* BEGIN
* GithubCacheEntry.create
  * SELECT to check for existing object
  * INSERT
* Upload blob
* GithubCacheEntry#update (UPDATE)
* COMMIT

Now behavior is:

* SELECT to check for existing object
* Upload blob
* GithubCacheEntry.create
  * SELECT to check for existing object
  * INSERT

Previously, if the GithubCacheEntry#update failed, I believe
the uploaded object was never deleted from the blob store, but
that should only happen if the upload_id returned matched an
existing record, which seems unlikely without a bug in the
blob store.  This deletes the uploaded object the if the
INSERT fails.

I don't think the sleeping inside transaction is related to the H12
errors, but when increasing the Puma threads beyond the database
connection pool size, you definitely want to avoid sleeping inside
transactions.

Best reviewed without whitespace differences.