Add Clover#paginated_result helper method
This DRYs up the code the api routes used for listing objects. It
also uses typecast_params to prevent unexpected types in those
parameters from causing problems.  This is not strictly necessary,
as committee checks api requests, and already raises for invalid
types. However, the goal is to stop using params completely in
routes and helpers, and trying to allow it for api but disallow
it for web would result in unnecessary complexity.