Use typecast_params in billing info routes
These use str!, which will result in a 400 if the parameter is not
submitted or is not a string.  It's technically only needed on the
tax_id and email parameters (where there are string methods called
on the result), but since this is taking input from the web form,
it is expected that all parameters should be submitted, unless
the user is monkeying with the form before submitting it.