Use custom invalid value messages for typecast_params
This will allow simplifying future code.  It currently simplifies
the usage alert limit check.  Since typecast_params was already
handling the conversion, instead of returning nil on a failed
conversion, have it raise, and handle the failure in the error
handler.

Treat the typecast_params error similar to a validation error, so
it redirects back on failure for web requests.

This requires the latest Roda commit, so update to a Roda
checkout.