Upgrade acme-client gem
It looks like there is a new error type that needs to be handled
properly by acme-client "OrderNotReady". That is added to the gem in a
later version https://github.com/unixcharles/acme-client/pull/246 which
we did not upgrade to. Now, we have a page with this error for
certificate cec899231d861h3wrvfe8pbpj0. Therefore, I am upgrading the
client.