CertNexus Cert finalization idempotancy fix
We were setting the csr_key and finalizing the certificate. The problem
with that is, if the cert finalization endpoint errors out but it
essentially worked, we loose the csr_key information. To fix that, we
first record the key to the database and use that continuously to
finalize the certificate.