Use typecast_params when retrieving location from parameter
This is done in the web routes when creating objects.  Currently,
the location is passed as a uuid instead of a ubid.  Until that
is fixed, we need to handle it as a uuid.  Add a uuid type to
typecast_params and use that.  Also, only do a query if a valid
uuid format is given.