Make Account#suspend mark related payment methods as fraud in a single query
This previously had an 2*N+1 query issue, since it would load the
billing info in one query and update the payment methods in a
second query per project.  It's possible to eager load the billing
info to avoid one of the N+1 queries, but you would still have an
N+1 query issue for the payment method update query.

This approach does the update in a single query with a single
subquery. This actually avoids the billing_info table entirely,
using the fact that both projects and payment methods have a
foreign key pointing to it.