Rename ApiKey.create_with_id to ApiKey.create
The long term plan is to eliminate create_with_id.  ApiKey is the
only model that overrides create_with_id.  Change it to override
create, and move key from a local variable to an optional keyword
argument.

While here, do not allow "inference_endpoint" as the owner_table,
since inference api keys use "inference_endpoint" as the used_for,
and "project" as the owner_table.  Additionally, do not convert
the owner_table to a string, as it should already be passed as
a string.