Add node_exporter installation prog to VmHost
A new prog is added to install the node_exporter on VmHosts.

A flag is also added to make sure the application will serve the
metrics on localhost instead of all network interfaces.

For the VmHosts which are already setup and do not have the app
installed we can create this strand:

Strand.create(prog: "SetupNodeExporter", label: "start",
	stack: [{subject_id: vm_host.id}])