Simplify PrivateSubnet#connected_subnets
This results in the same plan, but is slightly simpler.

I also tried this approach:

```ruby
PrivateSubnet
  .where(id: DB[:connected_subnet].where(subnet_id_1: id).select(:subnet_id_2))
  .or(id: DB[:connected_subnet].where(subnet_id_2: id).select(:subnet_id_1))
  .all
```

In a development database, this resulted in significantly less planning
time (saving ~100us), and only modest increase in execution time
(increase of ~15us).  I expect in production scenarios, execution
time will outweigh planning timing, and we can use a prepared statement
to reduce planning time if we need to.