Create a page if the repository blob token will expire in 1 day
If the repository nexus is unable to refresh the blob storage token for
any reason, it may fail silently. We generate tokens that last for 5
days and refresh them when there are 2 days left. If the token can't be
refreshed with 1 day remaining, we create a page. The on-call engineer
should investigate the issue before it expires.