Emit a route exception for committee response errors
I'm not sure why we would want to not emit this error, as this
error should 100% of the time be the result of a bug in the
application or in openapi.yml. If committee accepts the request,
it should accept the application's response to the request.
Hiding these errors only complicates debugging.