Fix matcher for postgres resource patch endpoint
Previously, we didn't add `true` keyword, which caused the matcher to match
every patch request, without checking if there are more segments in the path
that needs to be matched. `true` keyword ensures that the matcher matches only
the exact path and not any sub-paths. It didn't cause any problem so far, as we
only had one patch endpoint.