Add edit inline button component
This commits adds a new button component for inline editing functionality.
When clicked, it converts the text into an editable input field, allowing
users to modify the content directly on the page. Currently it only supportss
text input fields, but can be extended to support other types of inputs such
as select dropdowns or radio buttons in the future.